refactor: Exception/Converter 구조 통일, completedToday 로직 단순화, situationalRecommendations N+1 수정
