+++ src/main/java/com/umc/halo/domain/content/storybook/controller/StorybookControllerDocs.java
@@ -1,6 +1,6 @@
 package com.umc.halo.domain.content.storybook.controller;
 
-import com.umc.halo.domain.content.storybook.apiPayload.HomeSuccessCode;
+import com.umc.halo.domain.content.storybook.exception.code.HomeSuccessCode;
 import com.umc.halo.domain.content.storybook.dto.response.HomeResponse;
 import com.umc.halo.domain.content.storybook.dto.response.StorybookDetailResponse;
 import com.umc.halo.domain.content.storybook.dto.response.StorybookListResponse;
