+++ src/main/java/com/umc/halo/domain/content/storybook/exception/code/HomeSuccessCode.java
@@ -1,4 +1,4 @@
-package com.umc.halo.domain.content.storybook.apiPayload;
+package com.umc.halo.domain.content.storybook.exception.code;
 
 import com.umc.halo.global.apiPayload.code.BaseSuccessCode;
 import lombok.Getter;
@@ -8,7 +8,6 @@ import org.springframework.http.HttpStatus;
 @Getter
 @RequiredArgsConstructor
 public enum HomeSuccessCode implements BaseSuccessCode {
-
     GET_HOME(HttpStatus.OK, "HOME200_1", "홈 화면을 성공적으로 조회했습니다.");
 
     private final HttpStatus status;
